move all logic to network attach
