+++ lib/worker-server.js
@@ -53,7 +53,6 @@ module.exports = new ponos.Server({
     'application.container.created': require('workers/application.container.created'),
     'application.container.died': require('workers/application.container.died'),
     'application.container.errored': require('workers/application.container.errored'),
-    'application.container.started': require('workers/application.container.started'),
     'build.container.created': require('workers/build.container.created'),
     'build.container.died': require('workers/build.container.died'),
     'build.container.started': require('workers/build.container.started'),
