+++ lib/workers/container.network.attached.js
@@ -53,7 +53,80 @@ module.exports.jobSchema = joi.object({
 module.exports.task = (job) => {
   const log = logger.child({ method: 'ContainerNetworkAttachedWorker' })
   log.trace('XXX container.network.attached')
-  return Instance.findOneByContainerIdAsync(job.id)
+  const containerId = job.id
+  const instanceId = job.inspectData.Config.Labels.instanceId
+  const contextVersionId = job.inspectData.Config.Labels.contextVersionId
+  const sessionUserGithubId = job.inspectData.Config.Labels.sessionUserGithubId
+
+  return Promise.resolve()
+    .then(() => {
+      log.trace('update query')
+      const updateData = {
+        dockerContainer: containerId,
+        dockerHost: job.host,
+        inspect: job.inspectData,
+        ports: keypather.get(job, 'inspectData.NetworkSettings.Ports')
+      }
+
+      log.trace({ updateData }, 'update query')
+      return Instance.markAsCreating(
+        instanceId,
+        contextVersionId,
+        containerId,
+        updateData
+      )
+    })
+    .then(() => {
+      log.trace('markAsStartingAsync')
+      return Instance.markAsStartingAsync(instanceId, containerId)
+    })
+    .then(() => {
+      log.trace('findOneByContainerIdAsync')
+      return Instance.findOneByContainerIdAsync(job.id)
+    })
+    .tap(workerUtils.assertFound(job, 'Instance'))
+    .then(function (instance) {
+      return InstanceService.modifyExistingContainerInspect(
+        instance._id,
+        job.id,
+        job.inspectData,
+        job.inspectData.NetworkSettings.IPAddress
+      )
+      .catch(function (err) {
+        const statusCode = keypather.get(err, 'output.statusCode')
+        if (statusCode === 409) {
+          const fatalError = new WorkerStopError('Instance not found', { job: job }, { level: 'info' })
+          throw fatalError
+        }
+        throw err
+      })
+    })
+    .then(function (instance) {
+      return Promise.try(function () {
+        if (instance.isolated) {
+          return Isolation.findOneAsync({_id: instance.isolated})
+            .then(function (isolation) {
+              log.trace({isolation: isolation}, 'found isolation to kill')
+              if (isolation && isolation.state === 'killing') {
+                return InstanceService.killInstance(instance)
+                  .return(true)
+              }
+            })
+        }
+      })
+        .then(function (instanceKilled) {
+          if (!instanceKilled) {
+            const payload = {
+              instance: instance.toJSON()
+            }
+            payload.instance._id = instance._id.toString()
+            return rabbitMQ.publishInstanceStarted(payload)
+          }
+        })
+    })
+    .then(() => {
+      return Instance.findOneByContainerIdAsync(job.id)
+    })
     .tap(workerUtils.assertFound(job, 'Instance'))
     .then(function (instance) {
       return InstanceService.modifyExistingContainerInspect(
